+++ src/main/java/com/FinalProject/dto/BookRequest.java
@@ -1,4 +1,4 @@
-package com.FinalProject.request;
+package com.FinalProject.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
