+++ src/main/java/com/FinalProject/dto/CategoryRequest.java
@@ -1,4 +1,4 @@
-package com.FinalProject.request;
+package com.FinalProject.dto;
 
 import lombok.*;
