+++ src/main/java/com/FinalProject/dto/StudentDto.java
@@ -1,16 +1,18 @@
-package com.FinalProject.dto.studentdto;
+package com.FinalProject.dto;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.NoArgsConstructor;
 
+@Builder
 @Setter
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class StudentRequest {
-    private Long ID;
+public class StudentDto {
+    private Long id;
     private String name;
     private String surname;
     private String faculty;
