+++ src/main/java/com/FinalProject/dto/StudentRequest.java
@@ -1,22 +1,19 @@
-package com.FinalProject.dto.studentdto;
+package com.FinalProject.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 import lombok.NoArgsConstructor;
-//import org.mapstruct.Mapper;
+import lombok.Setter;
 
-//@Mapper
-@Builder
 @Setter
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class UpdateStudentDto {
-    private Long id;
+public class StudentRequest {
+    private Long ID;
     private String name;
     private String surname;
     private String faculty;
     private String studentFIN;
+    private boolean deleteStatus;
 }
