+++ src/main/java/com/FinalProject/dto/UpdateStudentDto.java
@@ -1,21 +1,22 @@
-package com.FinalProject.dto.studentdto;
+package com.FinalProject.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.NoArgsConstructor;
+//import org.mapstruct.Mapper;
 
+//@Mapper
 @Builder
 @Setter
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class StudentDto {
+public class UpdateStudentDto {
     private Long id;
     private String name;
     private String surname;
     private String faculty;
     private String studentFIN;
-    private boolean deleteStatus;
 }
