--- src/main/java/com/FinalProject/security/exception/ValidationErrorDto.java
@@ -1,15 +1,0 @@
-package com.FinalProject.security.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-@AllArgsConstructor
-public class ValidationErrorDto<T> {
-
-    private String message;
-    private T value;
-
-}
