+++ src/main/java/com/FinalProject/security/repository/UserRepository.java
@@ -10,9 +10,4 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
-
-
-
-
-
 }
