+++ src/main/java/com/FinalProject/service/AuthorService.java
@@ -2,7 +2,7 @@ package com.FinalProject.service;
 
 import com.FinalProject.dto.AuthorsDto;
 import com.FinalProject.dto.BookDto;
-import com.FinalProject.request.BookRequest;
+import com.FinalProject.dto.BookRequest;
 import com.FinalProject.model.Book;
 import org.springframework.stereotype.Service;
 
@@ -10,12 +10,7 @@ import java.util.List;
 
 @Service
 public interface AuthorService {
-//● List
-//● Create
-//● Update
-//● View
-//● Delete
-//● Show author books
+
 
     List<AuthorsDto> getAuthors();
