+++ src/main/resources/templates/navbar/sidebar.html
@@ -25,15 +25,12 @@
             <p th:text="${session.lastname}">Lastname</p>
         </div>
         <div class="middle-side-panel">
-            <div>
-                <b><a href="/">Home</a></b>
-            </div>
+
             <div>
                 <b><a href="/authors">Authors</a></b>
             </div>
             <div>
                 <b><a href="/book">Books</a></b>
-                <hr>
             </div>
 
             <div>
